Refactor VideoPlayer component to move utilities into their own file, and break out the xhr metadata fetching into its own function.
